+++ astropy/io/fits/tests/test_image.py
@@ -1793,6 +1793,7 @@ def test_bzero_implicit_casting_compressed():
     hdu = fits.open(filename)[1]
     hdu.data
 
+
 def test_bzero_mishandled_info(tmpdir):
     # Regression test for #5507:
     # Calling HDUList.info() on a dataset which applies a zeropoint
@@ -1801,7 +1802,7 @@ def test_bzero_mishandled_info(tmpdir):
     filename = tmpdir.join('floatimg_with_bzero.fits').strpath
     hdu = fits.ImageHDU(np.zeros((10, 10)))
     hdu.header['BZERO'] = 10
-    hdu.writeto(filename, clobber=True)
+    hdu.writeto(filename, overwrite=True)
     hdul = fits.open(filename)
     hdul.info()